fix: alternative line number message
We fix the format string for logs that expect line numbers from frames.
Since these can be set to `None`, and we use a custom string in this
case, we ensure that the format string uses a string placeholder to
avoid formatting error caused by the format string expecting a number
instead.